Add support for checking selected holds for different pick up locations. (#1987)
When editing multiple holds, they may have different pick up locations. When the situation is detected, editing of pickup locations is disabled. A message telling the user to edit an individual hold to change the pickup location is displayed instead. Since the check may be slow, a driver may indicate that only up to a configured limit of holds should be checked.